+++ previous-versions/files/src/sdk/models/operations/listfiles.ts
@@ -8,6 +8,9 @@ import { AxiosResponse } from "axios";
 import { Expose } from "class-transformer";
 
 export class ListFilesRequest extends SpeakeasyBase {
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 }
@@ -19,32 +22,53 @@ export class ListFilesRequest extends SpeakeasyBase {
  * This might be because your company or data connection id is wrong, or was already deleted.
  */
 export class ListFilesErrorMessage extends SpeakeasyBase {
+    /**
+     * `True` if the error occurred transiently and can be retried.
+     */
     @SpeakeasyMetadata()
     @Expose({ name: "canBeRetried" })
     canBeRetried?: string;
 
+    /**
+     * Unique identifier used to propagate to all downstream services and determine the source of the error.
+     */
     @SpeakeasyMetadata()
     @Expose({ name: "correlationId" })
     correlationId?: string;
 
+    /**
+     * Machine readable error code used to automate processes based on the code returned.
+     */
     @SpeakeasyMetadata()
     @Expose({ name: "detailedErrorCode" })
     detailedErrorCode?: number;
 
+    /**
+     * A brief description of the error.
+     */
     @SpeakeasyMetadata()
     @Expose({ name: "error" })
     error?: string;
 
+    /**
+     * Codat's service the returned the error.
+     */
     @SpeakeasyMetadata()
     @Expose({ name: "service" })
     service?: string;
 
+    /**
+     * The HTTP status code returned by the error.
+     */
     @SpeakeasyMetadata()
     @Expose({ name: "statusCode" })
     statusCode?: number;
 }
 
 export class ListFilesResponse extends SpeakeasyBase {
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -60,12 +84,18 @@ export class ListFilesResponse extends SpeakeasyBase {
     /**
      * Success
      */
-    @SpeakeasyMetadata({ elemType: shared.File })
-    files?: shared.File[];
+    @SpeakeasyMetadata()
+    files?: Record<string, any>[];
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;